src/function_app/requirements.txt
azure-functions==1.21.3
azure-cosmos==4.9.0
azure-identity==1.19.0
azure-search-documents==11.6.0
azure-storage-queue>=12.25
openai==1.57.4
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
pypdf==5.1.0
tiktoken==0.8.0
